// packages/editor/src/designable/designable-core/src/types.ts
export type IResizable = {
  width?: boolean
  height: boolean
  end?: (node: TreeNode, element: HTMLElement) => void
  move?: (node: TreeNode, element: HTMLElement, rect: IRect) => void
}

export type ITranslate = {
  x: boolean
  y: boolean
  reset: (node: TreeNode) => void
  end?: (node: TreeNode, diffX: string | number, diffY: string | number) => void
}